+++ src/@tests/functional/zanix-info.test.ts
@@ -1,6 +1,8 @@
 import { getAllZanixLibrariesInfo } from '../../modules/helpers/zanix/info.ts'
 import { versionRegex } from 'utils/regex.ts'
 import { assertMatch } from '@std/assert/assert-match'
+import { getZanixPaths } from 'modules/helpers/zanix/tree.ts'
+import { assert } from '@std/assert'
 
 Deno.test('Fetching Zanix lates version validation', async () => {
   const result = await getAllZanixLibrariesInfo()
@@ -13,4 +15,17 @@ Deno.test('Fetching Zanix lates version validation', async () => {
   assertMatch(result['@zanix/datamaster'].version, versionRegex)
   assertMatch(result['@zanix/server'].version, versionRegex)
   assertMatch(result['@zanix/tasker'].version, versionRegex)
+})
+
+Deno.test('getZanixPaths should return correct content from jsr', async () => {
+  const paths = getZanixPaths('library', '')
+
+  paths.subfolders.src.subfolders.utils.templates.base[0].PATH = '/src/utils/constants.ts'
+  const content = await paths.subfolders.src.subfolders.utils.templates.base[0].content({
+    metaUrl: import.meta.url,
+    relativePath: '../../../',
+  })
+
+  assert(content !== '')
+  assert(!content.includes('Not Found'))
 })