+++ src/modules/helpers/zanix/projects/library.ts
@@ -1,14 +1,20 @@
 import type { ZanixLibrarySrcTree } from 'typings/zanix.ts'
 
 import { getFolderName } from 'modules/helpers/paths.ts'
+import zanixLibInfo from '../info.ts'
 
 export const getLibraryFolders = (root: string): ZanixLibrarySrcTree => ({
   FOLDER: `${root}/src/modules`,
   get NAME() {
     return getFolderName(this.FOLDER)
   },
-  files: {
-    EXAMPLE: `${root}/src/modules/mod.ts`,
+  templates: {
+    base: [{
+      PATH: `${root}/src/modules/mod.ts`,
+      content(local) {
+        return zanixLibInfo.templateContent({ local, root, path: this.PATH })
+      },
+    }],
   },
   subfolders: {
     templates: {